% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{ask_tod}
\alias{ask_tod}
\title{Ask data to Trentino Open Data!}
\usage{
ask_tod(query, pack_sel = NULL, res_sel = NULL, rows = 100, ...)
}
\arguments{
\item{query}{A text input to search and download data}

\item{pack_sel}{Package selection. Custom parameter to override, force or set default package}

\item{res_sel}{Resource selection. Custom parameter to override, force or set default resource}

\item{rows}{Number of rows to be returned from the query}

\item{...}{Additional parameters to further customize behaviour}
}
\value{
A dataset if everything goes ok
}
\description{
dati.trentino.it is one of the most complete open data portal in europe,
and now you can easily discover and use
}
\examples{
\dontrun{
dat <- ask_tod('nati')
dat <- ask_tod('nati', pack_sel = 1)
dat <- ask_tod('dipendenti', pack_sel = 24, res_sel = 2)
}

}

