% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{download_resource}
\alias{download_resource}
\title{Handle the Resource download# Should return a data.frame}
\usage{
download_resource(res, sep)
}
\arguments{
\item{res}{A Resource data.frame}

\item{sep}{a custom separator, if present}
}
\value{
The actual data
}
\description{
Check the type of Resource and donwload&parse accordingly
}

