% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{get_spatial_data_frame}
\alias{get_spatial_data_frame}
\title{Extract shape file from remote zip.}
\usage{
get_spatial_data_frame(url)
}
\arguments{
\item{url}{The url of the zip file}
}
\value{
A SpatialDataFrame of selected resource.
}
\description{
Extract shape file from remote zip.
}

