% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{get_zip}
\alias{get_zip}
\title{Download zip from remote and unzip}
\usage{
get_zip(url)
}
\arguments{
\item{url}{The url of the zip file}
}
\value{
The path where files have been extracted
}
\description{
It create a temporary folder, download the selected resource
as zip file into the folder, and then extract all file into this path.
}

