% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{getSpatialDataFrame}
\alias{getSpatialDataFrame}
\title{Extract shape file from remote zip.}
\usage{
getSpatialDataFrame(url)
}
\arguments{
\item{url}{The url of the zip file}
}
\value{
A SpatialDataFrame of selected resource.
}
\description{
Extract shape file from remote zip.
}

