% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{menu_packages}
\alias{menu_packages}
\title{Packages Menu}
\usage{
menu_packages(s, rows = 100, pack_sel = NULL)
}
\arguments{
\item{s}{query}

\item{rows}{Number of rows to be returned in the menu, DEFAULT=100 (TODO: Possibily add a warning/message if rows is reached)}

\item{pack_sel}{Override Package selection}
}
\value{
The selected Package
}
\description{
Interactive menu to download a package
}

