% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{menu_resources}
\alias{menu_resources}
\title{Resources Menu}
\usage{
menu_resources(pack = NULL)
}
\arguments{
\item{pack}{A Package data.frame}
}
\description{
Interactive menu to download a Resource given a Package
}

