% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{organization_default}
\alias{organization_default}
\title{Organization default resource}
\usage{
organization_default(pack)
}
\arguments{
\item{pack}{Package df as returned by menu_packages}
}
\value{
The automatically selected Resource
}
\description{
If we know the structure of a organization tipical package, we can attempt to download it automatically
}

