% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{pack_default}
\alias{pack_default}
\title{Organization default resource}
\usage{
pack_default(pack, res)
}
\arguments{
\item{pack}{Package df as returned by menu_packages}

\item{res}{Resource options (RENAME - REFACTOR)}
}
\value{
The automatically selected Resource
}
\description{
If we know the structure of a organization tipical package, we can attempt to download it automatically
}

