% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{packages_menu}
\alias{packages_menu}
\title{Interactive menu to download a package}
\usage{
packages_menu(q = NULL, sel = NULL)
}
\arguments{
\item{q}{query}
}
\description{
Interactive menu to download a package
}

