% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{resources_menu}
\alias{resources_menu}
\title{Interactive menu to download a resource}
\usage{
resources_menu(resources = NULL)
}
\arguments{
\item{resources}{A data.frame as returned by \code{packages_menu}}
}
\description{
Interactive menu to download a resource
}

