% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{search_query}
\alias{search_query}
\title{Main search function}
\usage{
search_query(q = NULL, ...)
}
\arguments{
\item{q}{query}

\item{...}{further arguments}
}
\value{

}
\description{
This is the main search function and will return the result of the query
from dati.trentino
}
\examples{

ll <- search_query('nati')





}

