% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{select_resource}
\alias{select_resource}
\title{Handle the resource selection}
\usage{
select_resource(package, res_sel = NULL)
}
\arguments{
\item{package}{A Package data.frame}

\item{res_sel}{Override Resource selection}
}
\value{
NOTHING [TBC]
}
\description{
Search if the package have a default resource selection, if not ask and THEN download it
}

