% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{trentino}
\alias{trentino}
\title{Main function / entry point}
\usage{
trentino(q, rows = 100, pack_sel = NULL, res_sel = NULL)
}
\arguments{
\item{q}{A text input to search and download data}

\item{rows}{Number of rows to be returned from the query, DEFAULT=100 (TODO: Possibily add a warning/message when max is reached)}

\item{pack_sel}{Package selection. Custom parameter to override, force or set default package}

\item{res_sel}{Resource selection. Custom parameter to override, force or set default resource}
}
\value{
A dataset if everything goes ok, otherwise...(TODO)
}
\description{
This is the main function and then only that should be invoked, at least for now
}

