% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{ask_tod}
\alias{ask_tod}
\title{Main function / entry point}
\usage{
ask_tod(q, pack_sel = NULL, res_sel = NULL, rows = 100, ...)
}
\arguments{
\item{q}{A text input to search and download data}

\item{pack_sel}{Package selection. Custom parameter to override, force or set default package}

\item{res_sel}{Resource selection. Custom parameter to override, force or set default resource}

\item{rows}{Number of rows to be returned from the query}

\item{...}{Additional parameters to further customize behaviour}
}
\value{
A dataset if everything goes ok, otherwise...(TODO)
}
\description{
This is the main function and then only that should be invoked, at least for now
}

