% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{search_query}
\alias{search_query}
\title{Search function}
\usage{
search_query(q, rows = 100)
}
\arguments{
\item{q}{query}

\item{rows}{Number of rows to be returned from the query, DEFAULT=100 (TODO: Possibily add a warning/message when max is reached)}
}
\value{
A data.frame as returned from ckanr::package_search
}
\description{
This function search the portal for the specified query and return a data.frame containing the related packages
}

